Move PFTP-related Setup to a submodule of SoftwreSetup
